move shared helpers to extra
